+++ CHANGELOG.md
@@ -13,6 +13,10 @@ as of version 2.1.1.
 ### Changed
 - Changed the build system from poetry-core to setuptools.
 - Changed the VCPCode type to use class attributes instead of property functions.
+- Changed enums from `enum.Enum` to `enum.IntEnum`:
+  - `ColorPreset`
+  - `PowerMode`
+  - `InputSource`
 
 ### Removed
 - Removed support for end-of-life python version 3.8.