+++ CHANGELOG.md
@@ -6,19 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html),
 as of version 2.1.1.
 
-## [Unreleased]
+## [4.0.0] - 2025-05-26
 ### Added
 - Added additional logging for Windows API calls.
 
 ### Changed
 - Changed the build system from poetry-core to setuptools.
-- Changed the VCPCode type to use class attributes instead of property functions.
+- Changed the `VCPCode` type to use class attributes instead of property functions.
 - Changed `get_input_source` to return an `int` instead of an `InputSource`.
   - This simplifies code for users with USB Type-C monitors.
 - Changed enums from `enum.Enum` to `enum.IntEnum`:
   - `ColorPreset`
   - `PowerMode`
   - `InputSource`
+
+### Fixed
+- Fixed usability issues with mirrored displays in Windows.
 
 ### Removed
 - Removed support for end-of-life python version 3.8.
@@ -112,7 +115,8 @@ as of version 2.1.1.
 - Added a changelog.
 
 
-[Unreleased]: https://github.com/newAM/monitorcontrol/compare/3.1.0...HEAD
+[Unreleased]: https://github.com/newAM/monitorcontrol/compare/4.0.0...HEAD
+[4.0.0]: https://github.com/newAM/monitorcontrol/compare/3.1.0...4.0.0
 [3.1.0]: https://github.com/newAM/monitorcontrol/compare/3.0.3...3.1.0
 [3.0.3]: https://github.com/newAM/monitorcontrol/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/newAM/monitorcontrol/compare/3.0.1...3.0.2